+++ CHANGELOG.md
@@ -8,6 +8,10 @@ For older changes see the [archived Singularity change log](https://github.com/a
 ## v1.5.x changes
 
 - Update minimum go version to 1.25.3.
+- Add `--no-env` action and instance option and corresponding
+  `APPTAINER_NOENV` environment variable that can provide a
+  comma-separated list of environment variables to skip importing from
+  the host environment into the container.
 - Record image digest metadata (sha256 from `RepoDigests`), for OCI registry images.
   Also add the image name (ref) of the image from "docker", with registry and tag.
   This is useful for traceability, when using `docker.io` or a tag like `latest`.